# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SEPLAG API Rest",
        "description": "API Rest criada para o processo seletivo da SEPLAG 2025.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza a autenticação na api",
                "description": "Favor inserir os dados de email e senha (password)",
                "operationId": "ff999332b4fde4d1418d55fff84f397c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "senha"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "teste@teste.com.br"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha@123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "exp": {
                                            "type": "integer"
                                        },
                                        "usuario": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "nome": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "E-mail não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Email não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/registro": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Cadastrar de novos usuários",
                "description": "Criar novo usuário para acesso",
                "operationId": "38e4611ea044dd6eac72db007fcccdcb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "senha"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Maria Aparecida da Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "teste@teste.com.br"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha@123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "auth_id": {
                                            "type": "integer"
                                        },
                                        "auth_nome": {
                                            "type": "string"
                                        },
                                        "auth_email": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Refresh token autenticação",
                "description": "Gerar um novo token usando o token existente válido",
                "operationId": "b1568a1acfb18650550948c848e4c217",
                "responses": {
                    "200": {
                        "description": "Token atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "exp": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado - Token inválido ou expirado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token inválido ou expirado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Desconectar Usuário",
                "description": "Sair do usuário autenticado",
                "operationId": "97fafed12839d420c75cf57408010951",
                "responses": {
                    "200": {
                        "description": "Desconectado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout realizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}